requirements.txt
pandas
scikit-learn
streamlit
joblib
google-generativeai
python-dotenv
matplotlib
